+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - basketball achievements
+- service for correcting prices
 
 ### Fixed
 - rendering week fantasy leagues